# Optimize `Bag`.