[Test Case 1 & 2](https://gist.github.com/andersio/5cc8fd27a8adccf70d53036505069db3#file-perftests-swift)
Same environment as in #124

#### Baseline: `master`
One Observer:    0.018 sec (10%)
Eight Observers: 0.053 sec (5%)

#### A custom iterator to mitigate the time spent in protocol witness table for calling the default implementations.
One Observer:    0.012 sec (10%)
Eight Observers: 0.030 sec (6%)

#### Replace `Array` with `ContiguousArray` to avoid `swift_bridgeObjectRetain` calls in Array subscript getter, despite `BagElement` being a Swift struct.
One Observer:    0.010 sec (12%)
Eight Observers: 0.020 sec (3%)

#### Replace the monotonic ID in `Bag` with `ObjectIdentifier`.
Apparently no effect on iteration performance.